+++ src/lib-components/Flexible/ImpactNumberCards.vue
@@ -1,6 +1,8 @@
 <script setup lang="ts">
+import { computed } from 'vue'
 import type { PropType } from 'vue'
 import type { FlexibleImpactNumberCards } from '@/types/flexible_types'
+import { useTheme } from '@/composables/useTheme'
 
 import ImpactNumberCard from '@/lib-components/ImpactNumberCard.vue'
 
@@ -10,10 +12,15 @@ const { block } = defineProps({
     default: () => { },
   },
 })
+
+const theme = useTheme()
+const classes = computed(() => {
+  return ['impact-number-cards', theme?.value || '']
+})
 </script>
 
 <template>
-  <div class="impact-number-cards">
+  <div :class="classes">
     <div class="section-header">
       <h3
         v-if="block.sectionTitle"
@@ -40,57 +47,6 @@ const { block } = defineProps({
 </template>
 
 <style lang="scss" scoped>
-.impact-number-cards {
-  max-width: $container-l-main + px;
-  margin: 0 auto;
-
-  .section-header {
-    margin-bottom: var(--space-xl);
-  }
-
-  .section-title {
-    @include step-3;
-    color: var(--color-primary-blue-03);
-    margin-bottom: var(--space-m);
-  }
-
-  .section-summary {
-    @include step-0;
-    color: var(--color-black);
-
-    :deep(p) {
-      margin: 0;
-    }
-  }
-
-  .impact-number-cards-list {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    justify-content: flex-start;
-    align-content: flex-start;
-    gap: 32px 16px;
-  }
-
-  @media #{$medium} {
-    .impact-number-cards-list {
-      align-items: flex-start;
-
-      .flexible-impact-number-card {
-        width: calc((100% - 16px) / 2);
-      }
-    }
-  }
-
-  @media #{$small} {
-    .impact-number-cards-list {
-      display: flex;
-      flex-direction: column;
-
-      .flexible-impact-number-card {
-        width: 100%;
-      }
-    }
-  }
-}
+@import "@/styles/default/_impact-number-cards.scss";
+@import "@/styles/ftva/_impact-number-cards.scss";
 </style>