+++ src/stories/BlockCallToAction.stories.js
@@ -123,3 +123,29 @@ FTVASlimCTALongText.args = {
   title: '',
   text: '<p>Cat ipsum dolor sit amet, check cat door for ambush 10 times before coming in or gnaw the corn cob, but woops must get rid run around house maybe it comes off woops left brown marks on floor human slave clean now. Kitty kitty thinking longingly about tuna brine yet stinky cat for chase ball of string but attack the child. Scream for no reason at 4 am. Hide head under blanket so no one can see more napping, more napping all the napping is exhausting and roll over and sun my belly.</p>',
 }
+
+const mockFTVALight = {
+  title: 'LIGHT FTVA Call to Action',
+  text: '<p>Cat ipsum dolor sit amet, check cat door for ambush 10 times before coming in or gnaw the corn cob, but woops must get rid run around house maybe it comes off woops left brown marks on floor human can clean now. Kitty kitty thinking longingly about tuna brine yet stinky cat for chase ball of string but attack the child. Scream for no reason at 4 am. Hide head under blanket so no one can see more napping, more napping all the napping is exhausting and roll over and sun my belly.</p>',
+  name: 'Cat ipsum dolor',
+  to: '/help/foo/bar/',
+  isDark: false,
+}
+
+export const FTVA_LightWithButton = TemplateFTVA.bind({})
+FTVA_LightWithButton.args = {
+  ...mockFTVALight,
+}
+
+const mockFTVADark = {
+  title: 'DARK FTVA Call to Action',
+  text: '<p>Kitty kitty thinking longingly about tuna brine yet stinky cat for chase ball of string but attack the child. Scream for no reason at 4 am. Hide head under blanket so no one can see more napping, more napping all the napping is exhausting and roll over and sun my belly. Cat ipsum dolor sit amet, check cat door for ambush 10 times before coming in or gnaw the corn cob, but woops must get rid run around house maybe it comes off woops left brown marks on floor human can clean now.</p>',
+  name: 'Kitty ipsum dolor',
+  to: '/help/foo/bar/',
+  isDark: true,
+}
+
+export const FTVA_DarkWithButton = TemplateFTVA.bind({})
+FTVA_DarkWithButton.args = {
+  ...mockFTVADark,
+}